Move h3-tile endpoints into their own router file
