+++ api/app/main.py
@@ -6,37 +6,12 @@ import numpy as np
 import rasterio
 import uvicorn
 from app import settings
-from app.db import get_async_session
-from app.models import (
-    Dataset,
-    DatasetCountRequest,
-    DatasetCountTile,
-    DatasetMetadataRequest,
-    DatasetRequest,
-    DatasetsByLocationRequest,
-    HealthCheck,
-    TifAsPngRequest,
-)
-from app.routers import filters
-from app.services import (
-    dataset_count_to_bytes,
-    get_dataset_count_tiles_async,
-    get_dataset_metadata_results,
-    img_to_data_url,
-)
-from app.sql.bounds_fill import FILL, FILL_RES2
-from app.sql.dataset_coverage import DATASET_COVERAGE
-from app.sql.datasets_by_location import (
-    DATASETS_BY_LOCATION,
-    LOCATION_FILL,
-    LOCATION_FILL_RES2,
-)
-from fastapi import Depends, FastAPI, Response
+from app.models import HealthCheck, TifAsPngRequest
+from app.routers import dataset, filters
+from app.services import img_to_data_url
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from rasterio.warp import Resampling, reproject
-from shapely import wkt
-from sqlalchemy import func, select, text
-from sqlalchemy.ext.asyncio import AsyncSession
 
 app = FastAPI(
     title=settings.project_name,
@@ -53,9 +28,8 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-app.include_router(
-    filters.router
-)
+app.include_router(dataset.router)
+app.include_router(filters.router)
 
 
 @app.get("/health_check", response_model=HealthCheck)
@@ -64,128 +38,6 @@ async def health_check():
         "name": settings.project_name,
         "version": settings.version,
     }
-
-
-@app.post("/dataset_metadata/{index}")
-async def get_dataset_metadata(
-    index: str,
-    payload: DatasetMetadataRequest,
-    session: AsyncSession = Depends(get_async_session),
-):
-    filters={}
-    if payload.source_org:
-        filters["source_org"] = payload.source_org
-    results = await get_dataset_metadata_results(session, target=index, filters=filters)
-    return [
-        dict(
-            row._mapping,
-            bbox=wkt.loads(row._mapping['bbox']).bounds
-        )
-        for row in results
-    ]
-
-
-@app.post("/dataset_counts/{z}/{x}/{y}")
-async def get_dataset_counts(
-    payload: DatasetCountRequest,
-    z: int,
-    x: int,
-    y: int,
-    session: AsyncSession = Depends(get_async_session),
-):
-    filters = {}
-    if payload.source_org:
-        filters["source_org"] = payload.source_org
-    location = payload.location
-    should_hit_cache = not (payload.ignore_cache or location or filters)
-    cached_tile = None
-    if should_hit_cache:
-        cached_tile = await session.execute(
-            select(DatasetCountTile.dataset_counts).where(
-                DatasetCountTile.z == z,
-                DatasetCountTile.x == x,
-                DatasetCountTile.y == y,
-            )
-        )
-    header_kwargs = {}
-    if cached_tile:
-        dataset_count_bytes = cached_tile.scalar()
-        header_kwargs = {"headers": {"X-Tile-Cache-Hit": "true"}}
-    else:
-        resolution = payload.resolution
-        results = await get_dataset_count_tiles_async(session, z, x, y, resolution, location, filters)
-        if payload.debug_json_response:
-            return [row._mapping for row in results]
-        dataset_counts = {'index': [], 'dataset_count': []}
-
-        for row in results:
-            mapping = row._mapping
-            dataset_counts['index'].append(mapping.index)
-            dataset_counts['dataset_count'].append(mapping.dataset_count)
-        dataset_count_bytes = dataset_count_to_bytes(dataset_counts)
-        if should_hit_cache:
-            async with session:
-                dataset_count_tile = DatasetCountTile(
-                    z=z,
-                    x=x,
-                    y=y,
-                    dataset_counts=dataset_count_bytes,
-                )
-                session.add(dataset_count_tile)
-                await session.commit()
-    return Response(
-        content=dataset_count_bytes, media_type="application/octet-stream", **header_kwargs
-    )
-
-
-@app.post("/dataset_coverage/{z}/{x}/{y}")
-async def get_dataset_coverage(
-    payload: DatasetRequest,
-    z: int,
-    x: int,
-    y: int,
-    session: AsyncSession = Depends(get_async_session),
-):
-    resolution = payload.resolution
-    location = payload.location
-    query = DATASET_COVERAGE.format(fill_query=FILL_RES2 if resolution == 2 else FILL)
-    query = text(query).bindparams(z=z, x=x, y=y, location=location, resolution=resolution, dataset_id=payload.dataset_id)
-    results = await session.execute(query)
-    return [row[0] for row in results.fetchall()]
-
-
-@app.post("/dataset_count/")
-async def get_dataset_count(
-    session: AsyncSession = Depends(get_async_session),
-):
-    result = await session.execute(select(func.count(Dataset.id)))
-    return {"dataset_count": result.scalar_one()}
-
-
-@app.post("/datasets_by_location/")
-async def get_datasets_by_location(
-    payload: DatasetsByLocationRequest,
-    session: AsyncSession = Depends(get_async_session),
-):
-    location = payload.location
-    resolution = payload.resolution
-    parents_array = ["fill_index"] + [
-        f"h3_cell_to_parent(fill_index, {res})" for res in range(0, resolution)
-    ]
-    parents_comma_delimited = ", ".join(parents_array)
-    query = DATASETS_BY_LOCATION.format(
-        parents_array=parents_comma_delimited,
-        fill_query=LOCATION_FILL_RES2 if resolution == 2 else LOCATION_FILL
-    )
-    query = text(query).bindparams(location=location, resolution=resolution)
-    results = await session.execute(query)
-    return [
-        dict(
-            row._mapping,
-            bbox=wkt.loads(row._mapping['bbox']).bounds
-        )
-        for row in results.fetchall()
-    ]
 
 
 @app.post("/tif_as_png/")
